chore: make temporary info page describing how to upload documents
